Fix TDX test failing due to graphics

+++ tests/virtualization/universal/waitfor_guests.pm
@@ -17,12 +17,27 @@ use version_utils 'is_sle';
 # Open vnc and take screenshot of guests
 sub screenshot_vnc_guests {
     foreach my $guest (keys %virt_autotest::common::guests) {
+        next if $guest =~ /-tdx/;
         # Wait for virt-viewer to fully connect and display VNC stream
         enter_cmd "virt-viewer -f $guest & sleep 21 && killall virt-viewer";
         sleep 19;
         record_info "$guest", "$guest screenshot";
         save_screenshot();
         sleep 6;
+    }
+}
+
+sub upload_serial_console_guests {
+    foreach my $guest (keys %virt_autotest::common::guests) {
+        my $log = "/tmp/${guest}-serial.log";
+        if (script_run("test -f '$log'") == 0) {
+            upload_logs($log, failok => 1);
+        } else {
+            record_info(
+                "Serial console",
+                "No serial log found for $guest"
+            );
+        }
     }
 }
 
@@ -112,11 +127,13 @@ sub run {
     }
     assert_script_run('virsh list --all');
     wait_still_screen 1;
+    upload_serial_console_guests();
 }
 
 sub post_fail_hook {
     my ($self) = @_;
     screenshot_vnc_guests();
+    upload_serial_console_guests();
     access_vm_profiles();
     collect_virt_system_logs();
     $self->SUPER::post_fail_hook;
